Bug 1683177 - Improve mochitest web server shutdown and avoid exception under py3 on Windows; r=ahal
The exception seen in this bug can be avoided by simply replacing the terminate() call
with kill(). While we're here, I've tried to simplify the shutdown logic and improve
logging.

Differential Revision: https://phabricator.services.mozilla.com/D107528

+++ testing/mochitest/runtests.py
@@ -567,21 +567,18 @@ class MochitestServer(object):
     def stop(self):
         try:
             with closing(urlopen(self.shutdownURL)) as c:
-                c.read()
-
-            # TODO: need ProcessHandler.poll()
-            # https://bugzilla.mozilla.org/show_bug.cgi?id=912285
-            #      rtncode = self._process.poll()
-            rtncode = self._process.proc.poll()
-            if rtncode is None:
-                # TODO: need ProcessHandler.terminate() and/or .send_signal()
-                # https://bugzilla.mozilla.org/show_bug.cgi?id=912285
-                # self._process.terminate()
-                self._process.proc.terminate()
+                self._log.info(six.ensure_text(c.read()))
         except Exception:
             self._log.info("Failed to stop web server on %s" % self.shutdownURL)
             traceback.print_exc()
-            self._process.kill()
+        finally:
+            if self._process is not None:
+                # Wait briefly for normal shutdown, kill if necessary.
+                try:
+                    self._process.wait(1)
+                except TimeoutException:
+                    self._process.kill()
+                    self._log.info("Web server killed.")
 
 
 class WebSocketServer(object):
@@ -630,7 +627,8 @@ class WebSocketServer(object):
         self._log.info("runtests.py | Websocket server pid: %d" % pid)
 
     def stop(self):
-        self._process.kill()
+        if self._process is not None:
+            self._process.kill()
 
 
 class SSLTunnel:
